fix(mobile): unify review metadata sections
