+++ waltid-applications/waltid-wallet-demo-compose/sharedUI/src/commonMain/kotlin/id/walt/walletdemo/compose/ui/WalletUiTestTags.kt
@@ -33,6 +33,7 @@ internal object WalletUiTestTags {
     val PresentationVerifier = tag("presentationVerifier")
     val PresentationCancelButton = tag("presentationCancelButton")
     val PresentationVerifierSection = tag("presentationVerifierSection")
+    val PresentationVerifierInformationSection = tag("presentationVerifierInformationSection")
     val PresentationResponseProtectionSection = tag("presentationResponseProtectionSection")
     val PresentationTechnicalDetailsSection = tag("presentationTechnicalDetailsSection")
     val VerifierTechnicalDetailsToggle = tag("verifierTechnicalDetailsToggle")
