+++ waltid-applications/waltid-wallet-demo-ios/iosApp/iosApp/Components/ClaimGroupView.swift
@@ -5,25 +5,15 @@ struct ClaimGroupView: View {
 
     var body: some View {
         if !group.items.isEmpty {
-            VStack(alignment: .leading, spacing: 8) {
-                Text(group.title.uppercased())
-                    .font(.caption.weight(.bold))
-                    .foregroundStyle(.tint)
-                    .accessibilityLabel(group.title)
-                    .accessibilityIdentifier(WalletAccessibilityID.claimGroup(group.title))
-
-                VStack(alignment: .leading, spacing: 12) {
-                    ForEach(Array(group.items.enumerated()), id: \.element.id) { index, item in
-                        if index > 0 {
-                            Divider()
-                        }
-                        ClaimValueRow(item: item)
+            ReviewMetadataSection(title: group.title) {
+                ForEach(Array(group.items.enumerated()), id: \.element.id) { index, item in
+                    if index > 0 {
+                        Divider()
                     }
+                    ClaimValueRow(item: item)
                 }
-                .padding()
-                .background(Color(.systemGray6))
-                .clipShape(RoundedRectangle(cornerRadius: 8))
             }
+            .accessibilityIdentifier(WalletAccessibilityID.claimGroup(group.title))
         }
     }
 }
