+++ waltid-applications/waltid-wallet-demo-ios/iosApp/iosApp/ViewModels/MockWalletClient.swift
@@ -139,8 +139,8 @@ actor MockWalletClient: WalletClient {
                         logoAltText: nil
                     ),
                     clientURI: "https://verifier.example",
-                    policyURI: nil,
-                    termsOfServiceURI: nil
+                    policyURI: "https://verifier.example/privacy",
+                    termsOfServiceURI: "https://verifier.example/terms"
                 )
             },
             responseURI: URL(string: "https://verifier.example/response"),
