+++ waltid-applications/waltid-wallet-demo-ios/iosApp/iosApp/WalletAccessibilityID.swift
@@ -29,6 +29,7 @@ enum WalletAccessibilityID {
     static let presentationVerifier = identifier("presentationVerifier")
     static let presentationCancelButton = identifier("presentationCancelButton")
     static let presentationVerifierSection = identifier("presentationVerifierSection")
+    static let presentationVerifierInformationSection = identifier("presentationVerifierInformationSection")
     static let presentationResponseProtectionSection = identifier("presentationResponseProtectionSection")
     static let presentationTechnicalDetailsSection = identifier("presentationTechnicalDetailsSection")
     static let verifierTechnicalDetailsToggle = identifier("verifierTechnicalDetailsToggle")
